+++ src/bpp/admin/seria_wydawnicza.py
@@ -1,13 +1,16 @@
-# -*- encoding: utf-8 -*-
+from ..models.seria_wydawnicza import Seria_Wydawnicza
+from .core import BaseBppAdminMixin
+
 from django.contrib import admin
 
-from .core import CommitedModelAdmin
-from ..models.seria_wydawnicza import Seria_Wydawnicza
 
-
-class Seria_WydawniczaAdmin(CommitedModelAdmin):
-    list_display = ['nazwa', ]
-    search_fields = ['nazwa', ]
+class Seria_WydawniczaAdmin(BaseBppAdminMixin, admin.ModelAdmin):
+    list_display = [
+        "nazwa",
+    ]
+    search_fields = [
+        "nazwa",
+    ]
 
 
 admin.site.register(Seria_Wydawnicza, Seria_WydawniczaAdmin)