+++ src/bpp/admin/wydawnictwo_ciagle.py
@@ -2,17 +2,19 @@ from dal import autocomplete
 from django import forms
 from django.forms.utils import flatatt
 from djangoql.admin import DjangoQLSearchMixin
+from import_export import resources
 from mptt.forms import TreeNodeChoiceField
 from taggit.forms import TextareaTagWidget
 
 from pbn_api.models import Publication
+from . import BaseBppAdminMixin
 from .actions import (
     ustaw_po_korekcie,
     ustaw_przed_korekta,
     ustaw_w_trakcie_korekty,
     wyslij_do_pbn,
 )
-from .core import CommitedModelAdmin, KolumnyZeSkrotamiMixin, generuj_inline_dla_autorow
+from .core import KolumnyZeSkrotamiMixin, generuj_inline_dla_autorow
 
 # Widget do automatycznego uzupełniania punktacji wydawnictwa ciągłego
 from .element_repozytorium import Element_RepozytoriumInline
@@ -20,6 +22,7 @@ from .grant import Grant_RekorduInline
 from .helpers import (
     MODEL_OPCJONALNIE_NIE_EKSPORTOWANY_DO_API_FIELDSET,
     MODEL_Z_OPLATA_ZA_PUBLIKACJE_FIELDSET,
+    EksportDanychMixin,
     OptionalPBNSaveMixin,
     sprawdz_duplikaty_www_doi,
 )
@@ -177,13 +180,41 @@ class Wydawnictwo_Ciagle_Zewnetrzna_Baza_DanychInline(admin.StackedInline):
     form = Wydawnictwo_Ciagle_Zewnetrzna_Baza_DanychForm
 
 
+class Wydawnictwo_CiagleResource(resources.ModelResource):
+    class Meta:
+        model = Wydawnictwo_Ciagle
+        exclude = [
+            "tekst_przed_pierwszym_autorem",
+            "tekst_po_ostatnim_autorze",
+            "search_index",
+            "tytul_oryginalny_sort",
+            "legacy_data",
+            "cached_punkty_dyscyplin",
+            "opis_bibliograficzny_cache",
+            "opis_bibliograficzny_autorzy_cache",
+            "opis_bibliograficzny_zapisani_autorzy_cache",
+            "slug",
+        ]
+        export_order = [
+            "id",
+            "tytul_oryginalny",
+            "tytul",
+            "rok",
+            "ostatnio_zmieniony",
+        ]
+
+
 class Wydawnictwo_CiagleAdmin(
     DjangoQLSearchMixin,
     OptionalPBNSaveMixin,
     KolumnyZeSkrotamiMixin,
     AdnotacjeZDatamiOrazPBNMixin,
-    CommitedModelAdmin,
+    BaseBppAdminMixin,
+    EksportDanychMixin,
+    admin.ModelAdmin,
 ):
+    resource_class = Wydawnictwo_CiagleResource
+
     djangoql_completion_enabled_by_default = False
     djangoql_completion = True