+++ src/bpp/admin/zrodlo.py
@@ -11,7 +11,7 @@ from ..models import (  # Publikacja_Habilitacyjna
     Punktacja_Zrodla,
     Zrodlo,
 )
-from .core import CommitedModelAdmin
+from .core import BaseBppAdminMixin
 from .filters import PBN_UID_IDObecnyFilter
 from .helpers import (
     ADNOTACJE_FIELDSET,
@@ -97,7 +97,9 @@ class ZrodloForm(forms.ModelForm):
         fields = "__all__"
 
 
-class ZrodloAdmin(DjangoQLSearchMixin, ZapiszZAdnotacjaMixin, CommitedModelAdmin):
+class ZrodloAdmin(
+    DjangoQLSearchMixin, ZapiszZAdnotacjaMixin, BaseBppAdminMixin, admin.ModelAdmin
+):
     djangoql_completion_enabled_by_default = False
     djangoql_completion = True